After the old git repo history got corrupted, starting afresh
Unfortunately git got corrupted. I tried to find the culprit and fix but
alas, no luck. Inspite of moving the corrupted files (hash'es), I gave
up because the corrupted object seemed to be a tree object. So starting
afresh here again. I have zipped up the old .git folder and added as
part of this commit. Just in case it is needed some time in future...

lesson learnt when working on git is to take backups atleast once a
day...

So here we go again...