Changes in preparation for rel 0.42 to GES
The following changes are in preparation of release 0.42 to Global Edge
Software. Some of the key changes/fixes are

1. In returned values from ice_session_get_valid_pairs(), h_media is
NULL and nominated field is not set. This is now fixed.

2. If only one component is offered by the remote for a media stream
(say RTP and no RTCP), then session does not move to COMPLETE state.
This has now been fixed.

3. Renamed types.h to stun_types.h

4. Fixed coverity warnings - 5 out of total 8. Out of remaining 3, 2 are
in the GES application and one warning is related to stack usage which
can be ignored for now.