Some more changes for ICE stack release 0.42 to GES
Some of the changes in this commit are

1. An instance of the hard coding of the IP address family type in
connectivity check layer has been removed. The IP address family is
taken from the context.

2. When sending a stun message out via callback, the ip address family
is also given as a parameter as patr of callback. This will help the
application not to store the ip address family type within it's context.

3. The values of the enumeration stun_inet_address_type_t have been
renamed.