Added support for refreshing a TURN allocation
Added support for refreshing a TURN allocation with the TURN server.
Once the turn session is allocated, a refresh timer is started and every
time this timer fires, the session is refreshed by sending a new
REFRESH request message.

Though with this commit, the timer is started and refresh is done by
sending REFRESH request but the refresh timing does not seem to be
right. Also refresh stops after a while. This needs to be debugged.