Phew! minor victory - valid conn check request sent out for ice full
Minor but sweet victory. With this commit, a proper connectivity check
request was sent out to the peer and the peer responded with a success
response. Authentication and message integrity was fine.

As such, this comit has minor changes. But the key one is chosing of the
password for hmac key generation based on whether the stun message being
sent out is a binding request or response. This change was done in
ice_api.c.