Support for adding of a valid pair once successful conn check is done
Once a connectivity check succeeds, the pair needs to be added to the
valid list. The same has now been added as part of this commit.

1. Adding of valid pair once conn check succeeds.

2. If a peer reflexive cand pair is found due to conn check, then the
same is also added to the list of local candidates.

3. If conn check fails and role conflict is detected, then the role of
the agent for the current session is modified accordingly.

still to go... next is

- changes in the conn check cand pair states once one of the conn check
  succeeds.