Fixed the silly nasty bug of using DATA method instead of SEND method
While sending data to the turn server to be relayed to the peer address,
the turn layer was using DATA method when sending the indication instead
of SEND method. This was resulting in the server dropping the messages.
Made the change and now the turn server is passing the packets back and
forth.

What a silly mistake!!!  Need to read the specs properly and cross
check...

As of now, DATA indication is still not being handled from the turn
server. This would be the next step. The tunneled data from this needs
to be extracted and then processed.