Fixed invalid read issue while searching the stun transaction table
When the transaction instances were created by the various stun usages
like turn, conn check layer etc, the max size of the stun transaction
table was specified by the usage. However, when searching the table the
code was looping for a fixed number of times which was #defined
irrespective of the size specified by the usage. This was leading to
invalid reads which was caught by valgrind.

This has now been fixed. Now the looping is done only till the size
specified by the stun usage.