Changes to TURN layer to pass the application data to the ice layer
When DATA indication is received from the TURN/relay server, the
encapsulated application data is uncovered and passed to the
application. Changes for the same have been added as part of this
commit.

However, as of now there is some calloc and memory copy that is
involved in this data path. This needs to be removed later to optimize
the performance.