Connectivity check for a relayed candidate is succeeding
For the current test setup (ice clients and the turn server on same
host), the connectivity check for a relayed candidate is succeeding.
This is in addition to the connectivity check for host candidate which
was succeeding earlier.

Now the next task would be to make changes in the candidate pair states
depending on the success/failure of these connectivity checks.