Fixed a nasty bug in platform timer implementation, cause of sigsegv
Fixed a nasty bug in the platform timer implementation. In the doubly
linked list implementation of timer, when a timer node was deleted the
pointer for the prev node for the next node was not being assigned. This
was resulting in unpredictable crashed in the sample application.

This has been fixed in this commit. Hopefully no more crashed atleast
due to the timers.