Fix to set the conn check session to NULL once the same is destroyed
For an ice candidate pair, when the state changes to either SUCCEEDED or
FAILED, then the associated conn check session needs to be set to NULL.
Otherwise it is possible that when some other cand pair creates a conn
check session and gets the same handle (due to memory alloc and free
issues), this can create issues in the conn check sequence. Hence made
changes to set the conn check session handle to NULL as soon as it is
destroyed for any candidate pair.