Code changes to complete the connectivity checks for a media per sec 5.8
Earlier when processing the candidate pairs for a checklist, only the
pairs in WAITING state were being considered. Now as per sec 5.8 of ICE
RFC 5245, when no more pairs are available in WAITING state, highest
priority FROZEN candidates are also now being considered as part of code
changes in this commit.

With these changes, the connectivity checks for a media checklist are
completed properly.

Next changes would be to conclude the connectivity checks which are yet
to be done.