Adding to triggered check queue, if answer not yet received
If the answer has not yet been received from the peer, then such
incoming connectivity checks are responded to, and then they are added
to the triggered check queue. These pairs which are added to the
triggered check queue will be processes later once the checklist timer
fires once the answer has been received.

Next step is processing of these pairs in triggered check queue every
time the checklist timer fires. This is still incomplete.