Phew! conn checks are happening perfectly for controlled role
When the ice session is created in controlled mode, then connectivity
checks are happening properly and all the candidate pairs are moving to
SUCCEEDED state.

If incoming conn checks are received before the sdp answer from the peer
is received, then the same are responded to but the stored source and
destination candidates are stored. Once the sdp answer is received, then
these are processed.

To do next ....

1. Add validate candidate pairs to the valid list and notify the ice
agent application.
2. Test controlling role.