Added support for nomination timer to enable nominating of cand pair
When the session is using regular nomination and the agent's role is
controlling, then the nominated pairs will have to be determined after a
certain criteria. Currently we have choosen to

1. start a timer (nomination timer) which will decide as to when we will
stop the connectivity checks and try to evaluate among the available
valid candidate pairs and nominate among them.

2. Always nominate the candidate pair with the highest priority.

Point 1 above has now been implemented. However, point 2 is still under
implementation as of this commit.

completing point 2 would be the next immediate task.