misc changes to move into nominating state
Some miscellaneous changes so that when the nominating timer fires and
if there are valid pairs for each of the components of the media, then
the media moves into nominating state. Before moving into nominating
state, one candidate pair is nominated for each component from the
available valid pairs and a connectivity check is sent out for each of
such candidate pair with the USE-CANDIDATE attribute set.

Now need to handle responses for the nominated candidate pair
connectivity checks.