Done! connectivity checks completing fine in CONTROLLING role
OK, finally the task is done. Now even in controlling role, connectivity
checks happen file, candidate pair are nominated and the media moves to
completed state. A big victory after a lot of turmoil and pain.

Some caution though, since we have not tested the peer reflexive
candidates yet since with the current development setup I am not sure
how the scenario can be simulated.

Some major changes to get this working in this commit

1. Consolidation of media candidate pair list and the valid candidate
pair list into one. As of last commit, the checks were happening fine
when in controlling role. But nomination was not completing because
2 separate lists were maintained per media - one candidate pair list
and one valid pair list. However, the nominated candidate pairs were
in the valid list and we sent out connectivity checks for candidate
pairs in the valid list and we processed the incoming connectivity
responses over candidate pairs in the candidate pair list (not over
valid list). So nomination check responses were never handled and so
media would never to move to completed state.

    This has now been fixed by consolidating both the lists. The valid
list is now integrated into the candidate pair list. We could have gone
ahead with separate lists but that would have required us to duplicate
some function handler code in multiple places for processing of conn
check responses in the media candidate pair list and valid list. So that
has now been avioded with this integration. Also it helps to reduce the
memory consumption per media context.

2. Debug logs for the candidate pairs have been enhanced to dump the
validation state as well as the nomination status per candidate pair.
This helped me debug a very tricky situation where the media moved to
completed state as soon as the conn check response to the first
nominated candidate pair check response was received. This was because a
common flag was used in the candidate pair context for both nomination
status of a cand pair as well as selection of a cand pair for nomination
in controlling role. This has now been avoided by having separate flags
for both of them i.e., added a new flag 'check_nom_status' which will be
used to decide whether to add USE-CANDIDATE attribute in the
connectivity check request.

This commit has set us up for the first release of the ice full stack
code. However, thorough testing is required in all the possible modes
and some code cleanup as well.