Fix for nominating the pairs when both agents are ice-lite
This fix has already been shared with Global Edge. This fix is for a
scenario where our MB ice lite agent originates a call, the remote
party is also icelite and responds with a single host candidate for each
component. In this scenario, valid pairs were generated by the ICE stack
but were not nominated.

The fix given is to nominate the available single candidate pair in the
valid pair list.