Fix for session moving to COMPLETED state in case of use of rfc3484 - IPv6
When ice lite outgoing session is originated and suppose the remote is
also ice lite but has multiple IPv6 host candidates in answer, then our
ice lite stack nominates one of the pairs as per RFC 3484. Now the agent
sends another offer to peer with nominated candidate pair details and
when the peer responds, then the session state must move to COMPLETED
state. This was not happending earlier, but has now been fixed with this
commit.