Initial changes to support transient pair selections for aggressive alg
Initial preliminary code changes to support transient candidate pair
selection when the remote ice full agent uses aggressive nomination. In
this case, if the ice session has moved to completed state, then any new
nominated pair needs to be notified to the ice agent application.

The changes are incomplete as of this commit. The highest priority
nominated pair needs to be used.