Fix for stun re-transmission timer value calculation
Fix for stun re-transmission timer value calculation. The
re-transmission timer value was wrongly being computed due to which Rc
number of re-transmissions were being done but the timer value was
wrong. This is fixed in this commit. Now the overall transmission of
stun client request ends at 39.5 secs with recommended default values.