Multiple fixes for choosing the nominated pair in aggressive mode
Following fixes are part of this commit. These fixes are basically when
the remote is ICE FULL and is using an aggressive mode of nomination.

1. Suppose multiple connectivity check requests are received for the
same component in succession before receiving of conn check request for
the other component, then only one pair needs to be nominated - the one
with the highest candidate pair priority. This has been fixed in this
commit.

2. Handling of connectivity check requests even after media moves to
COMPLETED state. This can happen because of un-predictable nature of
Internet where higher priority connectivity check requests can arrive at
a later time than the lower priority ones.

3. Adding of candidate pair to valid list properly for a peer reflexive
candidate discovered for peer agent.