Added support for notifying the app when nominated pair is changed
Once the media moves to COMPLETED state, it is possible that a higher
priority ccandidate pair connectivity check arrives which results in the
change of the nominated candidate pair for a media. This is notified by
the miscellaneous event callback to the application. The support for
notifying the application when such an event happens is added in this
commit.