Fixes for incomplete/malformed/corrupted msgs when decoding - incomplete
Fixes for handling the incomplete/corrupted/malformed messages when
decoding the received stun messages.

1. Checks are added to make sure we do not end up performing any invalid
 read by checking the received buffer length.
2. Lengths of the attributes are validated against max allowed values
OR the fixed values depending upon the attribute.
3. Added printing of the data properly for ice controlled/controlling
attributes. This was not printed properly so far.
4. replaced some memcpy with stun_memcpy - trivial change

Further, the work is still incomplete as of this commi. Validation needs
to be completed for all the attributes. As of this commit, only some
attributes are handled. Others are to be taken care of ....