Fix to ignore duplicate pairs being added to valid list for regular nom
Fix to ignore duplicate pairs being added to valid list for regular
nomination. This happens when during regular nomination when the
candidate pairs are nominated, the peer might send multiple
re-transmissions of the connectivity check request. In such a case the
ice lite agent might treat each of the incoming conn check requests as
an independent separate conn check request and add to the valid list
even though that pair might already be part of the valid list.

In this commit, added a check so that before adding a pair to the valid
list, it is made sure that such a pair does not already exist.