Fixed issues with connectivity checks via external STUN server.
Fixed issues with connectivity checks via external free STUN server.
This was tested using the external IP address - 216.146.46.59 possibly
server is located at viagenie.ca). There was an issue when our ICE stack
was in controlling role (regular nomination) and was not nominating the
candidate pairs. This has been fixed now.

Other changes as part of this commit are

- added printing of incoming messages in the ice demo sample
- printing of extended attributes when received in incoming message