Fixed an issue where the stun binding session terminates due to timeout
Fixed an issue where the stun binding session terminates due to timeout.
When the stun binding session timesout, then the api returns that the
session terminated but does not indicate the session (handle) which
terminated. This has been fixed in the api. Addition of this change in
the stun binding layer API, led to the bug where the ice session does
not get terminated when the gathering of candidates failed due to stun
binding session timeout. This has now been fixed.

Similarly ICE api has been modified to indicate the ice session handle
which gets terminated when a timer event fires and is injected into the
API.