Added changes/fixes that were locally provided to GES earlier
Added changes/fixes that were locally provided to GES earlier. The major
change is the comparision of host addresses. Earlier, a simple string
comparision was used which used to work fine with IPv4 but would be not
suitable for IPv6 addresses since IPv6 address can be represented in
multiple ways. So the host comparision has been fixed as part of this
release.