Added changes for proper handling of local and remote peer-reflex cands
Added changes for proper discovery and handling of local and remote peer
reflexive candidates. Peer reflexive candidates was not at all tested
earlier because of unavailability of the setup and inability to simulate
the scenario. This can now be simulated by hardcoding the relayed
address in the turnserver to some other address.

Using the above method, tested the simple scenarios. Our stack was able
to properly discover both local and remote peer reflexive candidates and
add the same to the candidate pair list for the connectivity checks.

feel content now. accomplished something today.