Fixed an invalid read case for an unsuccessful scenario
Fixed an invalid read case for an unsuccessful scenario. While using
TURN server, if the ice session is destroyed when in between
connectivity check procedure, then the timers started by connectivity
check module are not handled when the media is in FAILED state. This has
now been fixed by handling the timer event.