Fixed bugs reported by Global Edge and also integrated changes suggested
Fixed bugs reported by Global Edge and also integrated changes suggested
by them. The 2 bugs reported by GLE and fixed as part of this release
are

1. Nov 08, 2011 - The nominated pair does not have the xor mapped
address as the local address, but instead has the base candidate

2. Nov 24, 2011 - Nomination connectivity check is not done for the RTCP
component. This bug does not seem to be an issue on our stack end but
seems likely to be an issue on the Global Edge part. But anyway, this
issue is tested with the current code base and this issue is not seen.

In addition to the above 2 bugs, also integrated some of the changes
suggested by Global Edge in their mails.

It is in good shape for release 0.54 to Global Edge Software.