MINDBRICKS ICE STACK release 0.54 to Global Edge Software.
MINDBRICKS ICE STACK release 0.54 to Global Edge Software. This
incremental release contains fixes for 2 bugs reported by Global Edge
and some other code changes. Among the 2 bugs reported, one does not
seem to be a bug in ICE stack. Anyway, this issue is not seen in the
latest code base, so did not try to reproduce with 0.53 release.

This is likely to be the last incermental release of ICE stack from
Mindricks to Global Edge Software. The plan is to release the final
version of ICE stack sometime next week, with any bug fixes for any bugs
if reported by Global Edge against this release 0.54.