Implemented support for requirements RFC4346 & RFC4360 for GES release
As part of final release to Global Edge Software, all requirements as
identified by the Checklist need to be implemented. As part of the same,
in this commit, the following two requirements from ICE RFC/draft are
implemented.

1. RFC4346 - If a relayed candidateis identical to a host candidate
(which can happen in rare cases),the relayed candidate MUST be
discarded.

2. RFC4360 - An agent MUST choose a set of candidates, one for each
component ofeach in-use media stream, to be default.