Implemented support for 5 more ice requirements for GES release
As part of final release to Global Edge Software, all requirements
as identified by the Checklist need to be implemented. As part of the
same, in this commit, the following 5 requirements from ICE RFC/draft
are implemented.

1. RFC4402 - If the requesthad contained the ICE-CONTROLLED attribute,
the agent MUST switch to the controlling role if it has not already done
so.

2. RFC4403 - If the requesthad contained the ICE-CONTROLLING attribute,
the agent MUST switch to the controlled role if it has not already done
so.

3. RFC4404 - Once it hasswitched, the agent MUST enqueue the candidate
pair whose checkgenerated the 487 into the triggered check queue.

4. RFC4405 - Note, however, that the tie-breaker value MUST NOT be
reselected.

5. RFC4400 - The content of either attribute MUST be the tie breaker
that was determined in Section 5.2. => No code change was required in
this commit to meet this requirement.