Implemented support for 10 more ice requirements for GES release
As part of final release to Global Edge Software, all requirements
as identified by the Checklist need to be implemented. As part of
the same, in this commit, the following 10 requirements from ICE
RFC/draft are implemented. The checklist is taken from the PacketCable
2.0 project that GES is implementing and this checklist is from
CableLabs.

1. RFC4522 - If neither ICE-CONTROLLING or ICE-CONTROLLED are present in
the request, the peer agent may have implemented a previous version of
this specification.

2. RFC4523 - If the agent is in the controlling role, and the ICE-CONTROLLING
attribute is present in the request:

3a RFC4524 - If the agent's tie-breaker is larger than or equal to
the contents of the ICE-CONTROLLING attribute, the agent generates a
Binding Error Response and includes an ERROR-CODE attributewith a value
of 487 (Role Conflict) but retains its role.

4. RFC4525 - If the agent's tie-breaker is less than the contents of
the ICE-CONTROLLING attribute, the agent switches to the controlled role.

5. RFC4526 - If the agent is in the controlled role, and the
ICE-CONTROLLED attribute is present in the request:

6. RFC4527 - If the agent's tie-breaker is larger than or equal to
the contents of the ICE-CONTROLLED attribute, the agent switches to the
controlling role.

7. RFC4528 - If the agent's tie-breaker is less than the contents of
the ICE-CONTROLLED attribute, the agent generates a Binding Error
Response and includes an ERROR-CODE attribute with a value of 487 (Role
Conflict) but retains its role.

8. RFC4529 - If the agent is in the controlled role and the
ICE-CONTROLLING attribute was present in the request, or the agent was in
the controlling role and the ICE-CONTROLLED attribute was present inthe
request, there is no conflict.

9. RFC4416 - An agent MUST examine the Binding Request for either the
ICE-CONTROLLING or ICE-CONTROLLED attribute.

10. RFC4417 - It MUST follow these procedures: