Added refresh functionality for stun binding session layer.
Added session refresh functionality for binding session layer. If
configured, the binding session will automatically keep on refreshing
the session by sending additional binding requests at provisioned time.
This was also to meet checklist requirement RFC4350 of ICE spec as
specified by PacketCable. Added the session refresh to the binding
session layer, so that this functionality can be usable to other modules
as well.

Following is the text of the checklist requirement

1. RFC4350 - For server reflexive candidates learned through aBinding
request, the bindings MUST be kept alive by additionalBinding Requests
to the server.