Fixed a bug where check list timer was not running when cp was added
Fixed a bug where check list timer was not running when a candidate pair
was added to triggered queue. In some scenarios, the check list timer
would be stopped if there is no pair left in the checklist for that
media. However, when a pair is added to the triggered queue, then the
timer must be started if stopped. This change has been done in this
commit.