Added changes to support dual-stack requirements for ICE-Lite
Global Edge had requested for dual-stack support for ICE-Lite though
support for dual-stack is not agreed upon in our license agreement. This
is an additional request that came from them via a mail dated
19/04/2012.

As part of this commit, changes have been made to support the same. This
was tested using a sample application. Also included changes to select
the default candidate in case of dual-stack scenarios by modifying the
API interface for ICE-Lite.

These changes have NOT been done in the ICE (ful) layer though.