Initial integration of the server application with the postgres database.
Integrated the server application with the postgres database.

1. When an allocation request is received, initial lookup is done in the
users table if an account exists for the received username in the TURN
allocation request.
2. When and if the allocation is approved, then a new row is added in
the allocation table for the newly allocation.

Still pending tasks
1. The current time is currently hardcoded. Need to find a way to get
the current time and populate it in the allocation table for some of the
timestamp fields such as allot_at, created_at, updated_at etc.
2. The IP address for the relayed address is currently hardcoded, needs
to be taken care of.
3. Currently the records/statistics in the user table are not updated
when the allocation request for a user is approved.
4. Not all validations are done for a received allocation request.