Fixed the issue of identifying the allocation table row uniquely.
This commit fixes a major issue of identifying a specific row in the
allocation table. When updating the dealloc_at column in the allocation
table when the user deallocates a context, we were earlier making use of
the allocation handles that was returned by the turns module. This
handle was stored in the table and was used for selecting the row upon
deallocation of a context by user. However, this handle is a memory
pointer for the allocation context that is returned by the turns module
and may not be unique all the times. For eg: if an allocation context is
allocated and de-allocated immediately and again an allocation is
requested, then the alloc handle might end up being same.

Hence in order to get around this issue, when allocation context is
approved, the server app will provide a blob handle to the turns module
and turns module will pass the blob for any further events that it
passes back to the server applciation. And the server makes use of the
allocation table row id as the blob, so upon any events when the turns
module notifies of any event, the server app can clearly pick the
allocation row.