More changes for SMP - use of shm for memory map, locks etc
More changes in order to support multiprocess concurrency support.

1) Added read-write lock for the list of allocations, so that this lock
is checked while adding, removing or traversing the list of allocations.
2) Lock for each allocation, so that this is acquired before doing any
operation on that particular allocation. Code changes are incomplete for
this yet (the lock needs to be acquired every time  before working on
the allocation)
3) Other minor changes such as changing the username in the allocation
context from pointer to static memory, etc.

More changes are required because of smp support

1) The instance is currently allocated on the heap. Once the fork()
happens, then the instance becomes a copy of the original in each forker
process, and they operate on a copy separately. Do we have a scenario
where all the processes need to operate on the same instance copy? In
such a case, the instance needs to be put on the shared memory.

2) The transaction handle request message is currently allocated on the
heap when the initial allocation request message is received. And the
decision whether to accept or reject the allocation is passed to the
user decision process which decides on the same and post a response back
to the main signaling processes. In case the decision process response
is handled by another process (than the one which handled the initial
request), then the handle to the initial request is invalid since the
pointer is no more valid in the current process. This becomes a problem
in forming a response and sending the allocation response since some of
the components from the initial request are required. One of the ways is
to allocate such requests on the shared memory either within the
allocation structure or a separate shared memory for such messages since
the duration for which they are required is quite less.

3) Currently transaction layer holds state but re-transmission is not
handled because timers are not properly implemented. For a TURN server,
it makes sense not to hold state but to treat each request as a new one,
calculate the response separately and respond appropriately. This would
require changes to transaction fsm state machine. Also this would
eliminate the other issue which is that - currently transaction list is
maintained in the heap and this will not work across multiple processes.
Also eliminates the need to do re-transmission and handling of timers
all together.