Support for sending allocation response subsequent to approval or rejection
The initial allocation request is currently allocated on the heap when the
initial allocation request message is received. And the decision whether to
accept or reject the allocation is passed to the user decision process which
decides on the same based on the database content and posts a response back
to the main signaling processes. In case the decision process response
is handled by another process (than the one which handled the initial
request), then the handle to the initial request is invalid since the
pointer is no more valid in the current process. This becomes a problem
in forming a response and sending the allocation response since some of
the components from the initial request are required.

This commit fixes the scenario mentioned above. In order to fix it, we
currrently maintain the raw stun buffer in the allocation context and
destroy the message handle when passing to the decision process. When
sending a response subsequent to approval, we again encode the raw
buffer to get the message handle. Using this new message handle, the
response is constructed and sent out.