Moving all the timer list in platform library into shared memory
This commit basically moves the timer list into a shared memory so that
all the worker processes can easily start and stop the timers. Also the
thread which checks the list to see if any timer has expired is now
moved into the master process. In case it finds any timer that has
expired, the event is posted to a socket that is being listed to by all
the worker processes. The worker processes now have a single thread
(earlier each worker process used to have it's own timer thread).

In addition to the above, there are some minor fixes.