Fixing the bug related to closing of shared socket descriptors
When multiple worker processes are used, when a socket descriptor is
closed then the same needs to be notified to the other worker processes
as well. This was not happening earlier, but has been fixed in this
commit. This has been done by properly making use of the recvmsg() and
sendmsg() socket api calls.

Now when a turn allocation is closed, the associated socket descriptor
for the relay port is closed not only in the notified process but also
in all the other worker processes who have handle to the same socket
descriptor.