Fixing all locking related issues with multiple processes
Fixing all locking related issues with the synchronization with multiple
processes. The major issue was that some of these synchronization
methods like posix mutexes and read-write locks were not set up to be
shared across multiple processes (PTHREAD_PROCESS_SHARED) due to which
we were seeing locking of processes on the shared items. For timers,
currently the read-write locks have been replaced by mutex. Now that the
real reason for blocking has been found, we should think of using the
read-write locks back in the future.

With these changes, the server is running fine without any locking
issues and also the sockers are being released properly by all the
worker processes. The CPU consumption is also quite less.