Fix for spurious fd events when using epoll in child processes
When epoll is used instead of select, and the epoll fd is created in the
parent process and then shared with the child processes, spurious events
are delivered to the child processes because the same kernel fd is
shared across all processes and any of the process will get events for
all the fds set by all the processes for the common epoll fd maintained
in kernel. To avoid these spurious events, only in case of epoll, the
epoll fd is created individually in all the worker processes after the
fork. This commit contains these changes and other changes to remove
some of the compilation warnings.