Updating the allocation columns for user in db
Earlier when an allocation was done, the relevant columns in the user
record were not incremented. When a new allocation is approved, the
allocs and the active_allocs columns need to be incremented so that the
admin policies can be enforced. Similarly, when an allocation is
terminated, the active_allocs column should be decremented. These
changes are part of this commit.