Validating of quota parameter when approving an incoming allocare req
When an allocation request is received, validations have been added in
this commit to check the following values against the provisioned ones
and if all conditions are met, then only the allocation is approved.
- allocations count
- active allocations count
- bandwidth used

Note: Also need to validate the 'requested_transport' attribute when an
incoming allocation request is received. However, that needs changes in
the turnserver web application as well, which need to be done first
before adding the validation here in this turn server.