initial changes for supporting dual authentication - incomplete
This commit contains initial changes for supporting dual authentication,
but the support is incomplete yet. Dual authentication is a means to
support TURN authentication in the TURN server for both

1] Traditional long-term authentication as specified in the TURN RFC
5766. Here the username and password are pre-set and the same are used
for authentication of TURN allocations.

2] Ephemeral credentials as specified in the TURN REST API draft
specification. Here the username and password are ephemeral in nature
and are created and shared with the clients using HTTP REST API.